+++ databento/historical/api/metadata.py
@@ -302,14 +302,14 @@ class MetadataHttpAPI(BentoHttpAPI):
             If an integer is passed, then this represents nanoseconds since UNIX epoch.
         symbols : List[Union[str, int]] or str, optional
             The product symbols to filter for. Takes up to 2,000 symbols per request.
-            If 'ALL_SYMBOLS' or ``None`` then will be for **all** symbols.
+            If 'ALL_SYMBOLS' or `None` then will be for **all** symbols.
         schema : Schema or str {'mbo', 'mbp-1', 'mbp-10', 'trades', 'tbbo', 'ohlcv-1s', 'ohlcv-1m', 'ohlcv-1h', 'ohlcv-1d', 'definition', 'statistics', 'status'}, default 'trades'  # noqa
             The data record schema for the request.
         stype_in : SType or str, default 'native'
             The input symbology type to resolve from.
         limit : int, optional
             The maximum number of records to include in the query.
-            If ``None`` then no limit.
+            If `None` then no limit.
 
         Returns
         -------
@@ -368,13 +368,13 @@ class MetadataHttpAPI(BentoHttpAPI):
             If an integer is passed, then this represents nanoseconds since UNIX epoch.
         symbols : List[Union[str, int]] or str, optional
             The product symbols to filter for. Takes up to 2,000 symbols per request.
-            If 'ALL_SYMBOLS' or ``None`` then will be for **all** symbols.
+            If 'ALL_SYMBOLS' or `None` then will be for **all** symbols.
         schema : Schema or str {'mbo', 'mbp-1', 'mbp-10', 'trades', 'tbbo', 'ohlcv-1s', 'ohlcv-1m', 'ohlcv-1h', 'ohlcv-1d', 'definition', 'statistics', 'status'}, default 'trades'  # noqa
             The data record schema for the request.
         stype_in : SType or str, default 'native'
             The input symbology type to resolve from.
         limit : int, optional
-            The maximum number of records to include in the query. If ``None`` then no limit.
+            The maximum number of records to include in the query. If `None` then no limit.
 
         Returns
         -------
@@ -436,13 +436,13 @@ class MetadataHttpAPI(BentoHttpAPI):
             The data feed mode for the request.
         symbols : List[Union[str, int]] or str, optional
             The product symbols to filter for. Takes up to 2,000 symbols per request.
-            If 'ALL_SYMBOLS' or ``None`` then will be for **all** symbols.
+            If 'ALL_SYMBOLS' or `None` then will be for **all** symbols.
         schema : Schema or str {'mbo', 'mbp-1', 'mbp-10', 'trades', 'tbbo', 'ohlcv-1s', 'ohlcv-1m', 'ohlcv-1h', 'ohlcv-1d', 'definition', 'statistics', 'status'}, default 'trades'  # noqa
             The data record schema for the request.
         stype_in : SType or str, default 'native'
             The input symbology type to resolve from.
         limit : int, optional
-            The maximum number of records to include in the query. If ``None`` then no limit.
+            The maximum number of records to include in the query. If `None` then no limit.
 
         Returns
         -------