+++ tests/test_historical_batch.py
@@ -129,3 +129,66 @@ class TestHistoricalBatch:
         ]
         assert call["timeout"] == (100, 100)
         assert isinstance(call["auth"], requests.auth.HTTPBasicAuth)
+
+    @pytest.mark.skipif(sys.version_info < (3, 8), reason="incompatible mocking")
+    def test_batch_list_files_sends_expected_request(
+        self,
+        mocker: MockerFixture,
+    ) -> None:
+        # Arrange
+        mocked_get = mocker.patch("requests.get")
+        job_id = "GLBX-20220610-5DEFXVTMSM"
+
+        # Act
+        self.client.batch.list_files(job_id=job_id)
+
+        # Assert
+        call = mocked_get.call_args.kwargs
+        assert (
+            call["url"]
+            == f"https://hist.databento.com/v{db.API_VERSION}/batch.list_files"
+        )
+        assert sorted(call["headers"].keys()) == ["accept", "user-agent"]
+        assert call["headers"]["accept"] == "application/json"
+        assert all(
+            v in call["headers"]["user-agent"] for v in ("Databento/", "Python/")
+        )
+        assert call["params"] == [
+            ("job_id", job_id),
+        ]
+        assert call["timeout"] == (100, 100)
+        assert isinstance(call["auth"], requests.auth.HTTPBasicAuth)
+
+    @pytest.mark.skipif(sys.version_info < (3, 8), reason="incompatible mocking")
+    def test_batch_download_single_file_sends_expected_request(
+        self,
+        mocker: MockerFixture,
+    ) -> None:
+        # Arrange
+        mocked_get = mocker.patch("requests.get")
+        job_id = "GLBX-20220610-5DEFXVTMSM"
+        filename = "glbx-mdp3-20220610.mbo.csv.zst"
+
+        # Act
+        self.client.batch.download(
+            job_id=job_id,
+            output_dir="my_data",
+            filename_to_download=filename,
+        )
+
+        # Assert
+        call = mocked_get.call_args.kwargs
+        assert (
+            call["url"]
+            == f"https://hist.databento.com/v{db.API_VERSION}/batch.list_files"
+        )
+        assert sorted(call["headers"].keys()) == ["accept", "user-agent"]
+        assert call["headers"]["accept"] == "application/json"
+        assert all(
+            v in call["headers"]["user-agent"] for v in ("Databento/", "Python/")
+        )
+        assert call["params"] == [
+            ("job_id", job_id),
+        ]
+        assert call["timeout"] == (100, 100)
+        assert isinstance(call["auth"], requests.auth.HTTPBasicAuth)